requirements.txt
Django
Pillow
django-rest-framework
python-decouple
psycopg2-binary
